Tekrar eden kodları sil.
